\name{FreeCompetitions}
\alias{FreeCompetitions}
\title{
This function returns all competitions that are released as free data from StatsBomb.com.}
\description{
This function easily scrapes the free competitions available from StatsBomb.com. Most effectively, this function is used internally in the StatsBombFreeEvents() function in order to pull in all of the freely available StatsBomb Data.}
\usage{
FreeCompetitions()
}

\examples{
comps <- FreeCompetitions()
}

