\name{competitions}
\alias{competitions}
\title{
This function returns all competitions that StatsBomb has data for.}
\description{
This function easily scrapes the competitions available from the StatsBomb API.}
\usage{
competitions(username, password)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\details{
}
\value{
\item{df}{Returns a data frame with all competitions available.}}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
competitions(username, password)

\keyword{StatsBomb}% use one of  RShowDoc("KEYWORDS")
\keyword{Data}% __ONLY ONE__ keyword per line
\keyword{Competitions}% __ONLY ONE__ keyword per line
\keyword{API}% __ONLY ONE__ keyword per line
