\name{FreeCompetitions}
\alias{FreeCompetitions}
\title{
This function returns all competitions that are released as free data from StatsBomb.com.}
\description{
This function easily scrapes the free competitions available from StatsBomb.com. Most effectively, this function is used internally in the StatsBombFreeEvents() function in order to pull in all of the freely available StatsBomb Data.}
\usage{
FreeCompetitions()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\details{
}
\value{
\item{df}{Returns a data frame with all competitions available for free.}}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
FreeCompetitions()

\keyword{StatsBomb}% use one of  RShowDoc("KEYWORDS")
\keyword{Data}% __ONLY ONE__ keyword per line
\keyword{Competitions}% __ONLY ONE__ keyword per line
\keyword{Free}% __ONLY ONE__ keyword per line
